+++ src/pages/videos/videos.store.js
@@ -27,50 +27,37 @@ export const selectSelectedItem = ({ state }) => {
   return flatItems.find(item => item.id === state.selectedItemId);
 }
 
-export const toViewData = ({ state, props }, payload) => {
+export const toViewData = ({ state }) => {
   const flatItems = toFlatItems(state.videosData);
   const flatGroups = toFlatGroups(state.videosData);
 
   // Get selected item details
-  const selectedItem = state.selectedItemId ? 
+  const selectedItem = state.selectedItemId ?
     flatItems.find(item => item.id === state.selectedItemId) : null;
 
-  // Compute display values for selected item
-  const selectedItemDetails = selectedItem ? {
-    ...selectedItem,
-    typeDisplay: selectedItem.type === 'video' ? 'Video' : 'Folder',
-    displayFileType: selectedItem.fileType || (selectedItem.type === 'video' ? 'MP4' : null),
-    displayFileSize: selectedItem.fileSize ? formatFileSize(selectedItem.fileSize) : null,
-    fullPath: selectedItem.fullLabel || selectedItem.name || '',
-  } : null;
-
   // Transform selectedItem into detailPanel props
-  const detailTitle = selectedItemDetails ? 'Video Details' : null;
-  const detailFields = selectedItemDetails && selectedItemDetails.type === 'video' ? [
-    { type: 'image', fileId: selectedItemDetails.thumbnailFileId, width: 240, height: 135, editable: true, accept: 'video/*', eventType: 'video-file-selected' },
-    { type: 'text', label: 'Name', value: selectedItemDetails.name },
-    { type: 'text', label: 'Type', value: selectedItemDetails.typeDisplay },
-    { type: 'text', label: 'File Type', value: selectedItemDetails.displayFileType, show: !!selectedItemDetails.displayFileType },
-    { type: 'text', label: 'File Size', value: selectedItemDetails.displayFileSize, show: !!selectedItemDetails.displayFileSize },
-    { type: 'text', label: 'Path', value: selectedItemDetails.fullPath, size: 'sm' }
-  ] : selectedItemDetails ? [
-    { type: 'text', label: 'Name', value: selectedItemDetails.name },
-    { type: 'text', label: 'Type', value: selectedItemDetails.typeDisplay },
-    { type: 'text', label: 'Path', value: selectedItemDetails.fullPath, size: 'sm' }
-  ] : [];
+  let detailFields
+  if (selectedItem) {
+    detailFields = [
+      { type: 'image', fileId: selectedItem.thumbnailFileId, width: 240, height: 135, editable: true, accept: 'video/*', eventType: 'video-file-selected' },
+      { type: 'text', value: selectedItem.name },
+      { type: 'text', label: 'File Type', value: selectedItem.fileType },
+      { type: 'text', label: 'File Size', value: formatFileSize(selectedItem.fileSize) },
+    ];
+  }
+
   const detailEmptyMessage = 'No selection';
-  
+
   return {
     flatItems,
     flatGroups,
     resourceCategory: 'assets',
     selectedResourceId: 'videos',
-    repositoryTarget: 'videos',
     selectedItemId: state.selectedItemId,
-    selectedItem: selectedItemDetails,
-    detailTitle,
+    detailTitle: undefined,
     detailFields,
     detailEmptyMessage,
+    repositoryTarget: 'videos',
   };
 }
